Remove ERC20 balance check when collecting debit in settlement layer.
The collectToken operation in SettleLayer previously compared the delta of the token
balance of the dex before and after the ERC transferFrom call. If this value was less
than the requested transfer amount (i.e. the token is "fee-on-transfer"), then the
transaction would abort.

Removing this check saves approximately 3000 gas in the happy case. The downside is that
fee-on-transfer tokens will now go through, which allows for the possibility of corrupted
pools. However these corrupted pools have no impact on the safety or functionality of the
non-corrupted non-fee-on-transfer token pools. Additionally rebase tokens could always
have created corrupted pools even before this change.

Functionally it's a usability consideration. Without explicitly checking for fee-on-
transfer an unsuspecting user interacting with a fee-on-transfer token may mint a
position that they can't fully recover. However saving gas on non fee-on-transfer tokens
is most likely the larger consideration for most users. Ultimately the fee-on-transer
usability check is better done off-chain rather than on-chain.