Prevent long-range birthday attacks on knockout merkles by injecting blockhash entropy
The pre-existing Merkle proof for knockout history used the timestamp and pool's fee
accumulator. Theoritically both of these components could be manipulated to allow for
a long-range attack. Since the Merkle hash is 160-bits, a possible birhtday attack
collission could be carried out with $10 billion of hash power.

This update eliminates the ability of an attacker to utilize a long-range birthday
collission. It injects the leading 160 bits of the block hash into the Merkle hash, which
prevents a long-range attacker from being able to fully predict/manipulate the proof
ahead of time.