Correct unchecked return values.
Two places in the codebase where return values were ignored were corrected.

The first was in TradeMatcher.sol, where the liquidity delta value from calling
crossKnockout was not being applied to the updated curve state. Practically this
didn't have any impact, since crossKnockout always returns a 0 value, but if
code is ever upgraded to the point this wasn't the case this could have an impact.
Now the returned delta value is applied to the curve liquidity.

The second was in CrocPolicy, where the return value of protocolCmd() calls on
CrocSwapDex contract were ignored. Every implementation of protocolCmd() returns
an empty byte string array, and the return value wasn't being used in any way.
So this changes protocolCmd() to have no return value to avoid the risk of a future
upgrade returning a value in protocolCmd(), while CrocPolicy ignores it.