Prevents spurious reverts on overflow in swap fee calculation
The swap fee calculation requires estimating the effective swap quantity
on every step. This entails calculating the amount of flow it would take
to reach the limit price of the swap step, and comparing that to the remaining
quantity in the swap, with the lesser being used for fee calculation on that
step.

The problem with this approach is that when the flow to reach the limitPrice
exceeds type(uint128).max it causes an overflow, since the underlying CurveMath
calculations bound all token flows at 128 bits. This can happen when using extreme
values for the limit price argument (such as a user who just sets limit price to
the max numerical value). These reverts are unnecessary, since if the limit price
flows exceed 128 bits, by definition they're less than the swap quantity which is
a uint128 bit argument.

This commit changes the logic to calculate the limitFlow value in uint256 precision.
Instead of reverting when exceeding uint128.max, the fee calculation returns the swap
quantity, which is the lesser bounded quantity by definition.